code_samples/production_rag/02_caching_optimization/requirements.txt
# Core dependencies for cost optimization system
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
redis>=4.0.0
scikit-learn>=1.0.0
numpy>=1.20.0
python-dotenv>=1.0.0
requests>=2.25.0

# Optional: For full AI functionality (may require additional setup)
# torch>=2.0.0
# transformers>=4.30.0